Fixes and Update LoginController
